perf(frontend): defer noncritical authenticated work
